gpu-requirements.txt
numpy==1.13.1
scipy==0.19.1
future==0.16.0
mne==0.15
nibabel==2.2.0
ipython==5.3.0
matplotlib==2.0.2
tensorflow-gpu==1.3.0
scikit-learn==0.18.1
Pillow==4.2.1
tqdm==4.15.0
statistics==1.0.3.5
matplotlib==2.0.2
envisage==4.6.0
PySide==1.2.4
mayavi==4.5.0
pysurfer==0.8.0
